chart-vision/requirements.txt
# Chart Vision - Stock Chart Analysis with Computer Vision
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Data & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Model Training
scikit-learn>=1.3.0
albumentations>=1.3.0  # Image augmentation
tqdm>=4.65.0

# For YOLO approach (optional)
ultralytics>=8.0.0  # YOLOv8

# Chart data
polygon-api-client>=1.12.0  # Polygon.io for stock data
# mplfinance>=0.12.0  # Optional - we use matplotlib directly

# Labeling tool
# labelImg  # Install separately: pip install labelImg

# Jupyter for experiments
jupyter>=1.0.0
ipywidgets>=8.0.0
